requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
redis
pydantic
pydantic-settings
aioredis
fastapi-jwt-auth
cryptography
bcrypt
python-multipart
PyJWT
passlib
confluent_kafka
kafka
python-jose